tests/wpt/tests/html/canvas/offscreen/text/2d.text.measure.width.space.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.text.measure.width.space</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.measure.width.space</h1>
<p class="desc">Space characters are converted to U+0020 and NOT collapsed</p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var f = new FontFace("CanvasTest", "url('/fonts/CanvasTest.ttf')");
  f.load();
  document.fonts.add(f);
  await document.fonts.ready;
  ctx.font = '50px CanvasTest';
  _assertSame(ctx.measureText('A B').width, 150, "ctx.measureText('A B').width", "150");
  _assertSame(ctx.measureText('A  B').width, 200, "ctx.measureText('A  B').width", "200");
  _assertSame(ctx.measureText('A \x09\x0a\x0c\x0d  \x09\x0a\x0c\x0dB').width, 650, "ctx.measureText('A \\x09\\x0a\\x0c\\x0d  \\x09\\x0a\\x0c\\x0dB').width", "650");
  _assert(ctx.measureText('A \x0b B').width >= 200, "ctx.measureText('A \\x0b B').width >= 200");

  _assertSame(ctx.measureText(' AB').width, 150, "ctx.measureText(' AB').width", "150");
  _assertSame(ctx.measureText('AB ').width, 150, "ctx.measureText('AB ').width", "150");

}, "Space characters are converted to U+0020 and NOT collapsed");
</script>
